libflash/mbox-flash: Always close windows before opening a new window
The MBOX protocol states that if an open window command fails then all
open windows are closed. Currently, if an open window command fails
mbox-flash will erroneously assume that the previously open window is
still open.

The solution to this is to mark all windows as closed before issuing an
open window command and then on success we'll mark the new window as
open.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>