libflash/test: Add tests for mbox-flash
A first basic set of tests for mbox-flash. These tests do their testing
by stubbing out or otherwise replacing functions not in
libflash/mbox-flash.c. The stubbed out version of the function can then
be used to emulate a BMC mbox daemon talking to back to the code in
mbox-flash and it can ensure that there is some adherance to the
protocol and that from a blocklevel api point of view the world appears
sane.

This makes these tests simple to run and they have been integrated into `make
check`. The down side is that these tests rely on duplicated feature
incomplete BMC daemon behaviour. Therefore these tests are a strong
indicator of broken behavaiour but a very unreliable indicator of
correctness.

Full integration tests with a 'read' BMC daemon are probably beyond the
scope of this repository.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>